wrapper_openDriver: fix string management logic
string copies and cats should be done on a clean array so
that we don't have garbage when changing parts of the string

The libraries are subject to be opened within the Sandbox when
using ChromiumOS, so this dlopen calls should just return the
handle to the already opened library.  Using a bad string will
result in an error and the sandbox is locked so no open of any
library is possible.

Also it is responsibility of the caller of dlopen to call dlclose
as calling dlclose will cause the sandbox to avoid any future open
of the library in case of error

Change-Id: I417335f057fb0332beb6598ef2f6ff2a446deb0a
Signed-off-by: Daniel Charles <daniel.charles@intel.com>